% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.R
\name{ars}
\alias{ars}
\title{Adapative Rejection Sampling}
\usage{
ars(N, f, a = -Inf, b = Inf, DFUN = NA, step = 0.5, est_mod = 0)
}
\arguments{
\item{N}{sample size (or number of observations).}

\item{f}{function f(x) which is log-concave and is proporitional to the density you wish to sample from}

\item{a}{first starting point for which f(x) is defined. The default is Inf.}

\item{b}{second starting pointfor which f(x) is defined. The default is -Inf.}

\item{DFUN}{the derivative of f(x)}

\item{step}{value of bandwidth around est_mode used to initialize function. The default is 0.5.}

\item{est_mod}{estimated mode of f(x) after log if f(x) is unbounded. The default is 0.}
}
\value{
\code{ars} returns a list containing the following components:
\item{f}{f(x)}
\item{fprime}{the derivative of f(x)}
\item{sample}{a vector of length N containing sampled values}
\item{hit_rate}{a value giving the proportion of accepted sample values over all evaluated values}
}
\description{
Adapative rejection sampling from log-concave functions. See Gilks and Wild (1992) for more details.
}
\examples{
# Example 1: Sample from Standard Normal Distribution
library(ars)
 N <- 1000
 x <- seq(-5, 5, len = N)
 dnor <- function(x) {
   return((1/(sqrt(2*pi)))*exp(-(x^2)/2))
 }
 out.norm <- ars(N=N, f=dnor, a=-5, b=5)
 plot(density(out.norm$sample), type="l", col="dodgerblue", lwd=2,
      main="Normal distribution", xlab="")
 y <- dnorm(x)
 points(x, y, type = "l", lwd=2)
# Example 2 : Sample from Laplace Distribution
 x <- seq(-5, 5, len = N)
 dlaplace <- function(x, m = 0, s = 1) {
 return(exp(-abs(x-m)/s)/(2*s))
 }
out.lap <- ars(N=N, f= dlaplace, a=-5, b=5)
plot(density(out.lap$sample), type="l", col="dodgerblue", lwd=2,
    main="Laplace distribution", xlab="")
y <- dlaplace(x, m=0, s=1)
points(x, y, type = "l")
}
\author{
Code developed by Xinyue Zhou, Cenzhuo Yao, Katherine C. Ullman, and Lauren Catherine Ponisio, based on Gilks and Wild (1992).
}
\references{
Gilks, W.R., P. Wild. (1992) Adaptive Rejection Sampling for Gibbs Sampling, Applied Statistics 41:337–348.
}

